Move test in dedicated file
